fix more 64bit warnings
BUG=skia:

Review URL: https://codereview.chromium.org/147683003

git-svn-id: http://skia.googlecode.com/svn/trunk@13190 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gm/nonclosedpaths.cpp
@@ -71,7 +71,7 @@ protected:
         // 0(may use hairline rendering), 10(common case for stroke-style)
         // 40 and 50(>= geometry width/height, make the contour filled in fact)
         static const int kStrokeWidth[] = {0, 10, 40, 50};
-        size_t numWidths = SK_ARRAY_COUNT(kStrokeWidth);
+        int numWidths = SK_ARRAY_COUNT(kStrokeWidth);
 
         static const SkPaint::Style kStyle[] = {
             SkPaint::kStroke_Style, SkPaint::kStrokeAndFill_Style
@@ -98,7 +98,7 @@ protected:
             for (size_t style = 0; style < SK_ARRAY_COUNT(kStyle); ++style) {
                 for (size_t cap = 0; cap < SK_ARRAY_COUNT(kCap); ++cap) {
                     for (size_t join = 0; join < SK_ARRAY_COUNT(kJoin); ++join) {
-                        for (size_t width = 0; width < numWidths; ++width) {
+                        for (int width = 0; width < numWidths; ++width) {
                             canvas->save();
                             SetLocation(canvas, counter, SkPaint::kJoinCount * numWidths);
